move navigation logic into navigation class to fix arrow buttons
