+++ finn/track_data_views/views/tree_view/navigation_widget.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, NamedTuple
 
-import pandas as pd
+import numpy as np
 from qtpy.QtWidgets import (
     QGroupBox,
     QHBoxLayout,
@@ -14,31 +14,35 @@ if TYPE_CHECKING:
     from finn.track_data_views import NodeSelectionList
 
 
+class NameData(NamedTuple):
+    itree: int
+    itrack: int
+    icell: int
+
+
 class NavigationWidget(QWidget):
     def __init__(
         self,
-        track_df: pd.DataFrame,
-        lineage_df: pd.DataFrame,
-        view_direction: str,
+        lineages: list,
+        displayed_lineages: list,
         selected_nodes: NodeSelectionList,
-        feature: str,
+        selected_nodes_data: dict,
+        get_view_direction,
+        get_feature,
     ):
         """Widget for controlling navigation in the tree widget
 
         Args:
-            track_df (pd.DataFrame): The dataframe holding the track information
-            view_direction (str): The view direction of the tree widget. Options:
-                "vertical", "horizontal".
             selected_nodes (NodeSelectionList): The list of selected nodes.
-            feature (str): The feature currently being displayed
         """
 
         super().__init__()
-        self.track_df = track_df
-        self.lineage_df = lineage_df
-        self.view_direction = view_direction
+        self.lineages = lineages
+        self.displayed_lineages = displayed_lineages
         self.selected_nodes = selected_nodes
-        self.feature = feature
+        self.selected_nodes_data = selected_nodes_data
+        self.get_view_direction = get_view_direction
+        self.get_feature = get_feature
 
         navigation_box = QGroupBox("Navigation [\u2b05 \u27a1 \u2b06 \u2b07]")
         navigation_layout = QHBoxLayout()
@@ -47,10 +51,10 @@ class NavigationWidget(QWidget):
         up_button = QPushButton("\u2b06")
         down_button = QPushButton("\u2b07")
 
-        left_button.clicked.connect(lambda: self.move("left"))
-        right_button.clicked.connect(lambda: self.move("right"))
-        up_button.clicked.connect(lambda: self.move("up"))
-        down_button.clicked.connect(lambda: self.move("down"))
+        left_button.clicked.connect(lambda: self.move_left())
+        right_button.clicked.connect(lambda: self.move_right())
+        up_button.clicked.connect(lambda: self.move_up())
+        down_button.clicked.connect(lambda: self.move_down())
 
         navigation_layout.addWidget(left_button)
         navigation_layout.addWidget(right_button)
@@ -65,127 +69,280 @@ class NavigationWidget(QWidget):
 
         self.setLayout(layout)
 
-    def move(self, direction: str) -> None:
-        """Move in the given direction on the tree view. Will select the next
-        node in that direction, based on the orientation of the widget.
-
-        Args:
-            direction (str): The direction to move. Options: "up", "down",
-                "left", "right"
-        """
-        if len(self.selected_nodes) == 0:
-            return
-        node_id = self.selected_nodes[0]
-
-        if direction == "left":
-            if self.view_direction == "horizontal":
-                next_node = self.get_predecessor(node_id)
-            else:
-                next_node = self.get_next_track_node(
-                    self.track_df, node_id, forward=False
-                )
-        elif direction == "right":
-            if self.view_direction == "horizontal":
-                next_node = self.get_successor(node_id)
-            else:
-                next_node = self.get_next_track_node(self.track_df, node_id)
-        elif direction == "up":
-            if self.view_direction == "horizontal":
-                next_node = self.get_next_track_node(self.lineage_df, node_id)
-                if next_node is None:
-                    next_node = self.get_next_track_node(self.track_df, node_id)
-            else:
-                next_node = self.get_predecessor(node_id)
-        elif direction == "down":
-            if self.view_direction == "horizontal":
-                # try navigation within the current lineage_df first
-                next_node = self.get_next_track_node(
-                    self.lineage_df, node_id, forward=False
-                )
-                # if not found, look in the whole dataframe to enable jumping to the
-                # next node outside the current tree view content
-                if next_node is None:
-                    next_node = self.get_next_track_node(
-                        self.track_df, node_id, forward=False
-                    )
-            else:
-                next_node = self.get_successor(node_id)
+    def move_left(self) -> None:
+        if self.get_view_direction() == "horizontal":
+            self.select_prev_cell()
         else:
-            raise ValueError(
-                f"Direction must be one of 'left', 'right', 'up', 'down', got {direction}"
-            )
-        if next_node is not None:
-            self.selected_nodes.add(next_node)
+            if self.get_feature() == "tree":
+                self.select_prev_lineage()
+            else:
+                self.select_prev_feature()
 
-    def get_next_track_node(
-        self, df: pd.DataFrame, node_id: str, forward=True
-    ) -> str | None:
-        """Get the node at the same time point in an adjacent track.
-
-        Args:
-            df (pd.DataFrame): The dataframe to be used (full track_df or subset
-                lineage_df).
-            node_id (str): The current node ID to get the next from.
-            forward (bool, optional): If true, pick the next track (right/down).
-                Otherwise, pick the previous track (left/up). Defaults to True.
-        """
-        # Determine which axis to use for finding neighbors
-        axis_label = "area" if self.feature == "area" else "x_axis_pos"
-
-        if df.empty:
-            return None
-        node_data = df.loc[df["node_id"] == node_id]
-        if node_data.empty:
-            return None
-
-        # Fetch the axis value for the given node ID
-        axis_label_value = node_data[axis_label].iloc[0]
-        t = node_data["t"].iloc[0]
-
-        if forward:
-            neighbors = df.loc[(df[axis_label] > axis_label_value) & (df["t"] == t)]
+    def move_right(self) -> None:
+        if self.get_view_direction() == "horizontal":
+            self.select_next_cell()
         else:
-            neighbors = df.loc[(df[axis_label] < axis_label_value) & (df["t"] == t)]
-        if not neighbors.empty:
-            # Find the closest index label
-            closest_index_label = (
-                (neighbors[axis_label] - axis_label_value).abs().idxmin()
+            if self.get_feature() == "tree":
+                self.select_next_lineage()
+            else:
+                self.select_next_feature()
+
+    def move_up(self) -> None:
+        if self.get_view_direction() == "vertical":
+            self.select_prev_cell()
+        else:
+            if self.get_feature() == "tree":
+                self.select_next_lineage()
+            else:
+                self.select_next_feature()
+
+    def move_down(self) -> None:
+        if self.get_view_direction() == "vertical":
+            self.select_next_cell()
+        else:
+            if self.get_feature() == "tree":
+                self.select_prev_lineage()
+            else:
+                self.select_prev_feature()
+
+    def select_next_cell(self):
+        """adds the cell at the next timepoint in the track to selected_nodes"""
+        node = self.selected_nodes[-1]
+        nd, vi = self.selected_nodes_data[node]
+        track = self.lineages[nd.itree][nd.itrack]
+        icell = vi + nd.icell + 1
+        if icell < len(track):
+            if icell == len(track) - 1:
+                iy = icell
+                vi = 0
+            else:
+                iy = 1
+                vi = icell - 1
+            newnode = self.lineages[nd.itree][nd.itrack][icell].node
+            self.selected_nodes_data[newnode] = (NameData(nd.itree, nd.itrack, iy), vi)
+            self.selected_nodes.add(newnode, False)
+        elif track[-1].marker == "triangle":
+            iprev = nd.itrack - 1
+            while self.lineages[nd.itree][iprev][0].time - 1 != track[-1].time:
+                iprev -= 1
+            track = self.lineages[nd.itree][iprev]
+            newnode = track[0].node
+            self.selected_nodes_data[newnode] = (NameData(nd.itree, iprev, 0), 0)
+            self.selected_nodes.add(newnode, False)
+
+    def select_prev_cell(self):
+        """adds the cell at the previous timepoint in the track to selected_nodes"""
+        node = self.selected_nodes[-1]
+        nd, vi = self.selected_nodes_data[node]
+        track = self.lineages[nd.itree][nd.itrack]
+        icell = vi + nd.icell - 1
+        if icell >= 0:
+            if icell == 0:
+                iy = 0
+                vi = 0
+            else:
+                iy = 1
+                vi = icell - 1
+            newnode = self.lineages[nd.itree][nd.itrack][icell].node
+            self.selected_nodes_data[newnode] = (NameData(nd.itree, nd.itrack, iy), vi)
+            self.selected_nodes.add(newnode, False)
+        elif track[0].marker == "square":
+            iprev = nd.itrack - 1
+            while (
+                iprev >= 0 and self.lineages[nd.itree][iprev][-1].time + 1 > track[0].time
+            ):
+                iprev -= 1
+            if (
+                iprev >= 0
+                and self.lineages[nd.itree][iprev][-1].time + 1 == track[0].time
+            ):
+                iparent = iprev
+            else:
+                inext = nd.itrack + 1
+                while (
+                    inext < len(self.lineages[nd.itree])
+                    and self.lineages[nd.itree][inext][-1].time + 1 > track[0].time
+                ):
+                    inext += 1
+                if (
+                    inext < len(self.lineages[nd.itree])
+                    and self.lineages[nd.itree][inext][-1].time + 1 == track[0].time
+                ):
+                    iparent = inext
+                else:
+                    return
+            track = self.lineages[nd.itree][iparent]
+            newnode = track[-1].node
+            self.selected_nodes_data[newnode] = (
+                NameData(nd.itree, iparent, len(track) - 1),
+                0,
             )
-            neighbor = neighbors.loc[closest_index_label, "node_id"]
-            return neighbor
-        return None
+            self.selected_nodes.add(newnode, False)
 
-    def get_predecessor(self, node_id: str) -> str | None:
-        """Get the predecessor node of the given node_id
-
-        Args:
-            node_id (str): the node id to get the predecessor of
-
-        Returns:
-            str | None: THe node id of the predecessor, or none if no predecessor
-            is found
+    def select_next_lineage(self):
         """
-        parent_id = self.track_df.loc[
-            self.track_df["node_id"] == node_id, "parent_id"
-        ].values[0]
-        parent_row = self.track_df.loc[self.track_df["node_id"] == parent_id]
-        if not parent_row.empty:
-            return parent_row["node_id"].values[0]
-        return None
-
-    def get_successor(self, node_id: str) -> str | None:
-        """Get the successor node of the given node_id. If there are two children,
-        picks one arbitrarily.
-
-        Args:
-            node_id (str): the node id to get the successor of
-
-        Returns:
-            str | None: THe node id of the successor, or none if no successor
-            is found
+        adds the cell at the same timepoint in the right neighboring track
+        to selected_nodes
         """
-        children = self.track_df.loc[self.track_df["parent_id"] == node_id]
-        if not children.empty:
-            child = children.to_dict("records")[0]
-            return child["node_id"]
-        return None
+        node = self.selected_nodes[-1]
+        nd, vi = self.selected_nodes_data[node]
+        track = self.lineages[nd.itree][nd.itrack]
+        time = track[vi + nd.icell].time
+        itree, itrack = nd.itree, nd.itrack + 1
+        itree = self.displayed_lineages.index(self.lineages[itree])
+        if itrack == len(self.displayed_lineages[itree]):
+            itrack = 0
+            itree += 1
+        found = False
+        while itree < len(self.displayed_lineages) and itrack < len(
+            self.displayed_lineages[itree]
+        ):
+            times = [x.time for x in self.displayed_lineages[itree][itrack]]
+            if time in times:
+                icell = times.index(time)
+                found = True
+                break
+            itrack += 1
+            if itrack == len(self.displayed_lineages[itree]):
+                itrack = 0
+                itree += 1
+        if found:
+            track = self.displayed_lineages[itree][itrack]
+            if icell == 0:
+                iy = 0
+                vi = 0
+            elif icell == len(track) - 1:
+                iy = icell
+                vi = 0
+            else:
+                iy = 1
+                vi = icell - 1
+            newnode = track[vi + iy].node
+            itree = self.lineages.index(self.displayed_lineages[itree])
+            self.selected_nodes_data[newnode] = (NameData(itree, itrack, iy), vi)
+            self.selected_nodes.add(newnode, False)
+
+    def select_prev_lineage(self):
+        """
+        adds the cell at the same timepoint in the left neighboring track
+        to selected_nodes
+        """
+        node = self.selected_nodes[-1]
+        nd, vi = self.selected_nodes_data[node]
+        track = self.lineages[nd.itree][nd.itrack]
+        time = track[vi + nd.icell].time
+        itree, itrack = nd.itree, nd.itrack - 1
+        itree = self.displayed_lineages.index(self.lineages[itree])
+        if itrack == -1:
+            itree -= 1
+            itrack = len(self.displayed_lineages[itree]) - 1
+        found = False
+        while itree >= 0 and itrack >= 0:
+            times = [x.time for x in self.displayed_lineages[itree][itrack]]
+            if time in times:
+                icell = times.index(time)
+                found = True
+                break
+            itrack -= 1
+            if itrack == -1:
+                itree -= 1
+                itrack = len(self.displayed_lineages[itree]) - 1
+        if found:
+            track = self.displayed_lineages[itree][itrack]
+            if icell == 0:
+                iy = 0
+                vi = 0
+            elif icell == len(track) - 1:
+                iy = len(track) - 1
+                vi = 0
+            else:
+                iy = 1
+                vi = icell - 1
+            newnode = track[vi + iy].node
+            itree = self.lineages.index(self.displayed_lineages[itree])
+            self.selected_nodes_data[newnode] = (NameData(itree, itrack, iy), vi)
+            self.selected_nodes.add(newnode, False)
+
+    def select_next_feature(self):
+        """
+        adds the cell at the same timepoint in the above neighboring track
+        to selected_nodes
+        """
+        node = self.selected_nodes[-1]
+        nd, vi = self.selected_nodes_data[node]
+        track = self.lineages[nd.itree][nd.itrack]
+        icell = vi + nd.icell
+        time = track[icell].time
+        feature = track[icell].area
+        feature_next = np.inf
+        found = False
+        for itree in range(len(self.displayed_lineages)):
+            for itrack in range(len(self.displayed_lineages[itree])):
+                times = [x.time for x in self.displayed_lineages[itree][itrack]]
+                if time in times:
+                    icell = times.index(time)
+                    thisfeature = self.displayed_lineages[itree][itrack][icell].area
+                    if feature < thisfeature < feature_next:
+                        itree_next, itrack_next, icell_next = itree, itrack, icell
+                        feature_next = thisfeature
+                        found = True
+        if found:
+            track = self.displayed_lineages[itree_next][itrack_next]
+            if icell_next == 0:
+                iy = 0
+                vi = 0
+            elif icell_next == len(track) - 1:
+                iy = len(track) - 1
+                vi = 0
+            else:
+                iy = 1
+                vi = icell_next - 1
+            newnode = track[vi + iy].node
+            itree_next = self.lineages.index(self.displayed_lineages[itree_next])
+            self.selected_nodes_data[newnode] = (
+                NameData(itree_next, itrack_next, iy),
+                vi,
+            )
+            self.selected_nodes.add(newnode, False)
+
+    def select_prev_feature(self):
+        """
+        adds the cell at the same timepoint in the below neighboring track
+        to selected_nodes
+        """
+        node = self.selected_nodes[-1]
+        nd, vi = self.selected_nodes_data[node]
+        track = self.lineages[nd.itree][nd.itrack]
+        icell = vi + nd.icell
+        time = track[icell].time
+        feature = track[icell].area
+        feature_prev = 0
+        found = False
+        for itree in range(len(self.displayed_lineages)):
+            for itrack in range(len(self.displayed_lineages[itree])):
+                times = [x.time for x in self.displayed_lineages[itree][itrack]]
+                if time in times:
+                    icell = times.index(time)
+                    thisfeature = self.displayed_lineages[itree][itrack][icell].area
+                    if feature > thisfeature > feature_prev:
+                        itree_prev, itrack_prev, icell_prev = itree, itrack, icell
+                        feature_prev = thisfeature
+                        found = True
+        if found:
+            track = self.displayed_lineages[itree_prev][itrack_prev]
+            if icell_prev == 0:
+                iy = 0
+                vi = 0
+            elif icell_prev == len(track) - 1:
+                iy = len(track) - 1
+                vi = 0
+            else:
+                iy = 1
+                vi = icell_prev - 1
+            newnode = track[vi + iy].node
+            itree_prev = self.lineages.index(self.displayed_lineages[itree_prev])
+            self.selected_nodes_data[newnode] = (
+                NameData(itree_prev, itrack_prev, iy),
+                vi,
+            )
+            self.selected_nodes.add(newnode, False)
